cucumber: fixed the expended external data
Apparently the views for the ci-jruby.org Jenkin's instance has
changed.

This is just temporary, since it uses live data, and it'll change
constantly.  A better solution needs to be figured out.  Either just
checking for the header (and possibly one or more results) or by stubbing
the data returned to the gem.